Refactoring metadata + allow to get metadata of a file form cli (#217)
* add some bencmark to compare capnp and protobuf v3

Tests to compare capnp and protobuf v3
we compare both size of encoded struct and speed of encoding

* fix meta_test capnp benchmark

* add benchmark for capnp-packed and msgpack

* add protobuf benchmark with re-used buffer

* Refactor metadata

- Metadata is now a normal object, we only use capnp when encoding and
decoding from the metadata server
- Change the schema of the metadata. Now we have a list of chunks and
each chunks can have multiple shards see
https://github.com/zero-os/0-stor/issues/135#issuecomment-325274579

* replace MetaSize with a method that walk on all chunks and compute the total size

* change Meta.Epoch from uint64 to int

* check for error during capnp encoding

* use same type for Meta.Epoch as time.Now().UnixNano() returns

* add more pipe combinaison in client_test

* distribution : use same key for both metadata and 0-stor.

* always save metadata after file write

* allow to choose the key use when uploading file

* update documentation regarding metadata

* don't write metadata to etcd during the data processing

* use filepath instead of path

* rename shards to pieces in distribution, it's more explicit

+++ client/walk_test.go
@@ -103,9 +103,9 @@ func testWalk(t *testing.T, forward bool) {
 	// walk over it
 	var wrCh <-chan *WalkResult
 	if forward {
-		wrCh = cli.Walk(firstKey, uint64(startEpoch), uint64(endEpoch))
+		wrCh = cli.Walk(firstKey, startEpoch, endEpoch)
 	} else {
-		wrCh = cli.WalkBack(firstKey, uint64(startEpoch), uint64(endEpoch))
+		wrCh = cli.WalkBack(firstKey, startEpoch, endEpoch)
 	}
 
 	var i int
